iot_py_client/requirements.txt
pyserial ~= 3.4
azure-iot-device ~= 2.1